Added the user component to crontab
I added the user root as the one to be running the crontab job, I don't believe it's needed though I feel safer having it here. I haven't fully tested this yet so if anyone has any issues with the crontab not restarting the system every 24 hours, please submit an issue.